[🔨refactor] 버튼 비활성화 체크 함수 분리
