--- src/utils/checkCategoryDuplication.ts
@@ -1,9 +1,0 @@
-import { CategoryType } from '../types/menuMangementType';
-export const checkDuplicate = (categoryList: CategoryType[], categoryName: string) => {
-	let isDuplicate = false;
-	const categoryNameList = categoryList.map((category) => category.category);
-	if (categoryNameList.includes(categoryName)) {
-		isDuplicate = true;
-	}
-	return isDuplicate;
-};
